Move `createDirectories()` to commons
